ASGARD-1045 - dry run option for launch config mass delete

+++ grails-app/controllers/com/netflix/asgard/LaunchConfigurationController.groovy
@@ -17,6 +17,7 @@ package com.netflix.asgard
 
 import com.amazonaws.services.autoscaling.model.AutoScalingGroup
 import com.amazonaws.services.autoscaling.model.LaunchConfiguration
+import com.netflix.asgard.model.JanitorMode
 import com.netflix.grails.contextParam.ContextParam
 import grails.converters.JSON
 import grails.converters.XML
@@ -144,6 +145,7 @@ class LaunchConfigurationController {
     private String doMassDelete(UserContext userContext, int daysAgo) {
         Check.atLeast(10, daysAgo, 'daysAgo')
         DateTime cutOffDate = new DateTime().minusDays(daysAgo)
+        JanitorMode mode = params.mode ? JanitorMode.valueOf(params.mode) : JanitorMode.EXECUTE
         Collection<AutoScalingGroup> allGroups = awsAutoScalingService.getAutoScalingGroups(userContext)
         Collection<LaunchConfiguration> allConfigs = awsAutoScalingService.getLaunchConfigurations(userContext)
         Collection<LaunchConfiguration> oldUnusedConfigs = allConfigs.findAll { LaunchConfiguration lc ->
@@ -152,12 +154,18 @@ class LaunchConfigurationController {
             configIsOld && configNotInUse
         }
         DateTimeFormatter formatter = ISODateTimeFormat.date()
-        String message = "Deleting ${oldUnusedConfigs.size()} unused launch configs from before" +
+
+        String executeMessage = "Deleting ${oldUnusedConfigs.size()} unused launch configs from before" +
                 " ${formatter.print(cutOffDate)} \n"
+        String dryRunMessage = "Dry run mode. If executed, this job would delete ${oldUnusedConfigs.size()} unused" +
+                "launch configs from before ${formatter.print(cutOffDate)} \n"
+        String message = JanitorMode.EXECUTE == mode ? executeMessage : dryRunMessage
         oldUnusedConfigs.sort { it.createdTime }
         oldUnusedConfigs.each { LaunchConfiguration lc ->
             try {
-                awsAutoScalingService.deleteLaunchConfiguration(userContext, lc.launchConfigurationName)
+                if (mode == JanitorMode.EXECUTE) {
+                    awsAutoScalingService.deleteLaunchConfiguration(userContext, lc.launchConfigurationName)
+                }
                 message += "Deleted ${formatter.print(lc.createdTime.time)} ${lc.launchConfigurationName} \n"
             } catch (Exception e) {
                 message += "Could not delete Launch Configuration ${lc.launchConfigurationName}: ${e} \n"
